# Implement parser without backtracking