--- a/stack-7.8.yaml
+++ b/stack-7.8.yaml
@@ -3,6 +3,7 @@ packages:
 - '.'
 extra-deps:
 - bytestring-lexing-0.5.0.2
+- scanner-0.2
 flags:
   hedis:
     dev: true