--- a/stack-head.yaml
+++ b/stack-head.yaml
@@ -1,7 +1,8 @@
 resolver: lts-5.3
 packages:
 - '.'
-extra-deps: []
+extra-deps:
+- scanner-0.2
 flags:
   hedis:
     dev: true