@Yuras 2016-04-01
CI failure looks unrelated to the change, and I can't reproduce it locally. (I have no idea why it fails actually)