@k-bx 2016-04-01
Yeah, I suspect using more recent redis in test-server will fix this, it's a common error after we migrated to dockerized travis.