@Yuras 2016-04-02
I decided to release the parser as a separate package because I think it will be useful in other contexts: http://hackage.haskell.org/package/scanner
Please let me know what you think about integrating it into hedis.