@informatikr 2016-04-02
A nice performance improvement and, with the parser in a separate package, the code is basically the same as before. No doubt this should go into hedis.