@Yuras 2016-04-03
I implemented it. But we should probably wait for scanner to appear in stackage lts snapshot?