@k-bx 2016-04-04
@Yuras there's no reason to wait for lts snapshot before merging/releasing this. First of all, hedis is constraint in stackage already (because of some dependency from lts constraining it, I've made a PR), secondly – packages first get into nightly, and then they are included into lts at some point. So all we have to do is to get you into a nightly snapshot by making a PR on stackage github repo which adds you or someone else as a maintainer.