@Yuras 2016-04-04
@k-bx scanner is already in nightly: https://www.stackage.org/package/scanner

> is this ready to be merged now?

Probably not, this failure makes me worry: https://travis-ci.org/informatikr/hedis/jobs/120553050
(It is the `force error reply` test case, right?) I can't reproduce it, but it doesn't look like a random glitch. It may (or may not) indicate a bug in resupplying scanner with more input, so it fails only when the input is chunked in some very specific way. I'm going to investigate it more.