@k-bx 2016-04-04
@Yuras the one you've linked is a "pipelining" bug, it appears rarely in Travis and it sometimes appears already (in master code), so I wouldn't worry too much about it