@k-bx 2016-04-04
Check out [raw output](https://s3.amazonaws.com/archive.travis-ci.org/jobs/120553050/log.txt) to see this:

```
[2K[1Fforce error reply: [[92mOK[0m]

[2K[1Fpipelining: [[91mFailed[0m]


[2K[1Feval unused replies: [[92mOK[0m]
```